# PERF: use StringHasTable for strings