@jreback 2016-12-12
cc @mrocklin