@jreback 2016-12-14
cc @mrocklin

if you can give this a stress test (and see if it helps in your perf tests), when you have a chance