@mrocklin 2016-12-14
I *think* I've narrowed down my original problem to something else.
Working on isolating it further now.  Will try this after that issue is
removed.

On Wed, Dec 14, 2016 at 11:05 AM, Jeff Reback <notifications@github.com>
wrote:

> cc @mrocklin <https://github.com/mrocklin>
>
> if you can give this a stress test (and see if it helps in your perf
> tests), when you have a chance
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/14859#issuecomment-267074198>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AASszImRcGD1PokmZSZMPJFRbLyO1ToCks5rIBPLgaJpZM4LKF2s>
> .
>